+++ Quotations Board Backend/Models/ImpliedYield/TBillYield.cs
@@ -7,7 +7,7 @@ public class TBillYield
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public string Id { get; set; } = null!;
     public DateTime YieldDate { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     [ForeignKey("TBill")]
     public string TBillId { get; set; } = null!;
     public TBill TBill { get; set; } = null!;
